refactor: move utils into handlers module to avoid import issues
